# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "BFD",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/stammuser/get": {
      "post": {
        "summary": "Get Stamm Users",
        "tags": [
          "Stamm User"
        ],
        "operationId": "GetStammUsers",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetStammUsersRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/stammuser": {
      "get": {
        "summary": "Get Stamm User",
        "tags": [
          "Stamm User"
        ],
        "operationId": "GetStammUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/stammuser/update": {
      "post": {
        "summary": "Update Stamm User",
        "tags": [
          "Stamm User"
        ],
        "operationId": "UpdateStammUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateStammUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/stammuserproject/get": {
      "post": {
        "summary": "Get Stamm User Projects",
        "tags": [
          "Stamm User Project"
        ],
        "operationId": "GetStammUserProjects",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetStammUserProjectsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/stammuserproject": {
      "get": {
        "summary": "Get Stamm User Project",
        "tags": [
          "Stamm User Project"
        ],
        "operationId": "GetStammUserProject",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/stammuserproject/update": {
      "post": {
        "summary": "Update Stamm User Project",
        "tags": [
          "Stamm User Project"
        ],
        "operationId": "UpdateStammUserProject",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateStammUserProjectRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Address": {
      "title": "Address",
      "example": {
        "streetNo": "",
        "city": "",
        "zip": 0
      },
      "type": "object",
      "properties": {
        "streetNo": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "zip": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "streetNo",
        "city",
        "zip"
      ]
    },
...
    "GetStammUsersRequest": {
      "title": "GetStammUsersRequest",
      "example": {
        "sortBy": "",
        "order": "",
        "search": "",
        "page": 1,
        "limit": 10,
        "organizer": ""
      },
      "type": "object",
      "properties": {
        "sortBy": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "search": {
          "type": "string"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "organizer": {
          "type": "string"
        }
      },
      "required": [
        "sortBy",
        "order",
        "search",
        "page",
        "limit",
        "organizer"
      ]
    },
    "UpdateStammUserRequest": {
      "title": "UpdateStammUserRequest",
      "example": {
        "id": "",
        "company": "",
        "firstName": "",
        "lastName": "",
        "email": "",
        "phone": "",
        "address": {
          "streetNo": "",
          "city": "",
          "zip": 0
        },
        "organizer": "",
        "otherContact": [
          {
            "name": "",
            "email": "",
            "phone": ""
          }
        ]
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "$ref": "#/definitions/Address"
        },
        "organizer": {
          "type": "string"
        },
        "otherContact": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OtherContact1"
          }
        }
      },
      "required": [
        "id",
        "company",
        "firstName",
        "lastName",
        "email",
        "phone",
        "address",
        "organizer",
        "otherContact"
      ]
    },
    "OtherContact1": {
      "title": "OtherContact1",
      "example": {
        "name": "",
        "email": "",
        "phone": ""
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "email",
        "phone"
      ]
    },
...
    "GetStammUserProjectsRequest": {
      "title": "GetStammUserProjectsRequest",
      "example": {
        "sortBy": "",
        "order": "",
        "search": "",
        "page": 1,
        "limit": 10,
        "stammUser": ""
      },
      "type": "object",
      "properties": {
        "sortBy": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "search": {
          "type": "string"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "stammUser": {
          "type": "string"
        }
      },
      "required": [
        "sortBy",
        "order",
        "search",
        "page",
        "limit",
        "stammUser"
      ]
    },
    "UpdateStammUserProjectRequest": {
      "title": "UpdateStammUserProjectRequest",
      "example": {
        "id": "",
        "name": "",
        "phone": "",
        "address": {
          "streetNo": "",
          "city": "",
          "zip": 0
        },
        "stammUser": "",
        "otherContact": [
          {
            "name": "",
            "email": "",
            "phone": ""
          }
        ]
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "$ref": "#/definitions/Address"
        },
        "stammUser": {
          "type": "string"
        },
        "otherContact": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OtherContact1"
          }
        }
      },
      "required": [
        "id",
        "name",
        "phone",
        "address",
        "stammUser",
        "otherContact"
      ]
    },
...
  },
...
}